# backend/multimodal-ai-service/src/main/resources/openapi-multimodal-ai.yaml
openapi: 3.0.3
info:
  title: Multimodal Medical AI Service API
  description: |
    Comprehensive multimodal AI service that integrates text, images, vitals, and genomic data
    for advanced diagnosis and treatment planning in healthcare.
  version: 1.0.0
  contact:
    name: AutoMed Development Team
    email: dev@automed.com
servers:
  - url: http://localhost:8086
    description: Development server
  - url: https://api.automed.com/multimodal-ai
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/v1/multimodal-ai/synthetic-data:
    post:
      summary: Generate synthetic medical data
      description: |
        Generates synthetic data for training and testing AI models
        while maintaining privacy and medical accuracy.
      operationId: generateSyntheticData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyntheticDataRequest'
      responses:
        '200':
          description: Successful synthetic data generation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyntheticDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: []
...
  /api/v1/multimodal-ai/models:
    get:
      summary: Get available AI models
      description: Retrieves information about available AI models.
      operationId: getAvailableModels
      responses:
        '200':
          description: Successful model information retrieval
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AiModelInfo'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'